Add @fresha/openapi-codegen-test-utils package
