+++ .circleci/config.yml
@@ -51,6 +51,7 @@ jobs:
       - run: npm run -w @fresha/openapi-codegen-server-elixir --if-present build && exit
           0
       - run: npm run -w @fresha/openapi-codegen-cli --if-present build && exit 0
+      - run: npm run -w @fresha/openapi-codegen-test-utils --if-present build && exit 0
       - run: npm run -w @fresha/openapi-diff --if-present build && exit 0
       - run: npm run -w @fresha/openapi-lint --if-present build && exit 0
       - run: npm run -w @fresha/openapi-validator --if-present build && exit 0
