+++ packages/openapi-codegen-test-utils/.eslintrc
@@ -1,0 +1,7 @@
+{
+  "extends": ["@fresha"],
+  "parser": "@typescript-eslint/parser",
+  "parserOptions": {
+    "project": ["./tsconfig.eslint.json"]
+  }
+}
