+++ packages/openapi-codegen-test-utils/jest.config.js
@@ -1,0 +1,1 @@
+module.exports = require('@fresha/jest-config');
