+++ packages/openapi-codegen-test-utils/package.json
@@ -1,7 +1,9 @@
 {
-  "name": "@fresha/jest-config",
+  "name": "@fresha/openapi-codegen-test-utils",
   "version": "0.1.0",
-  "description": "Common Jest configuration for this project",
+  "description": "Testing utilites for OpenAPI codegens",
+  "main": "build/index.js",
+  "types": "build/index.d.ts",
   "scripts": {
     "build": "tsc",
     "build:watch": "tsc --watch",
@@ -17,6 +19,9 @@
     "test": "jest --passWithNoTests",
     "typecheck": "tsc --noEmit"
   },
+  "keywords": [
+    "JSON:API"
+  ],
   "author": "Fresha Engineering",
   "contributors": [
     {
@@ -32,16 +37,20 @@
       "url": "https://github.com/mykulyak"
     }
   ],
-  "main": "build/index.js",
-  "types": "build/index.d.ts",
+  "license": "MIT",
   "devDependencies": {
     "@fresha/eslint-config": "0.1.0",
     "@fresha/typescript-config": "0.1.0",
     "@types/jest": "^29.1.2",
-    "jest": "^29.2.0",
-    "jest-circus": "^29.2.0",
-    "jest-junit": "^14.0.1",
-    "ts-jest": "^29.0.3",
     "typescript": "^4.7.2"
+  },
+  "dependencies": {
+    "@fresha/api-tools-core": "0.3.0",
+    "@fresha/jest-config": "0.1.0",
+    "@fresha/openapi-model": "0.4.0",
+    "jest-matcher-utils": "^29.1.2",
+    "prettier": "^2.7.1",
+    "ts-morph": "^16.0.0",
+    "winston": "^3.8.2"
   }
 }
