+++ packages/openapi-codegen-test-utils/src/jestConfig.ts
@@ -1,0 +1,8 @@
+import path from 'path';
+
+import baseConfig from '@fresha/jest-config';
+
+export default {
+  ...baseConfig,
+  setupFilesAfterEnv: [path.join(__dirname, 'setupEnv.js')],
+};
