+++ packages/openapi-codegen-test-utils/src/setupEnv.ts
@@ -1,0 +1,3 @@
+import { createMatchers } from "./matchers";
+
+createMatchers();
